Adding Journey History (Challenge 12)
Added #history and #journey attribute readers. #history returns an array and #journey is created upon #touch_in to create a hash for each journey. The hash keys are the entry stations and the hash values are the exit stations. Refactored to add #start_journey and #end_journey helper methods to uphold the Single Responsibility Principle for #touch_in and #touch_out.